Covering possible int overflow when comparing Strings
Covering a possible int overflow that could occur from int multiplication if MAX_EDIT_DISTANCE_MEMORY was ever increased.

In it's current state, this is unlikely to happen due to the if statement above it filtering out the larger values. However, think it's still useful to "fix" preemptively as is a little bit of a ticking time-bomb if MAX_EDIT_DISTANCE_MEMORY or other depending code is changed in the future.

+++ spock-core/src/main/java/org/spockframework/runtime/ExpressionInfoValueRenderer.java
@@ -134,7 +134,7 @@ public class ExpressionInfoValueRenderer {
       return "false\nStrings too large to calculate edit distance.";
     } else {
       // Check if we can add some context
-      if ((end1 - commonStart + 20) * (end2 - commonStart + 20) < MAX_EDIT_DISTANCE_MEMORY){
+      if (((long) end1 - commonStart + 20) * (end2 - commonStart + 20) < MAX_EDIT_DISTANCE_MEMORY){
         commonStart = Math.max(0, commonStart - 10);
         end1 = Math.min(str1.length(), end1 + 10);
         end2 = Math.min(str2.length(), end2 + 10);
